+++ spec/support/config/rake.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+Rake.application.options.trace = false